refactor: refactor into colour grid component
